feat(repodb): move repodb logic inside meta directory under pkg
