+++ pkg/extensions/extension_search_disabled.go
@@ -9,8 +9,8 @@ import (
 
 	"zotregistry.io/zot/pkg/api/config"
 	"zotregistry.io/zot/pkg/log"
+	"zotregistry.io/zot/pkg/meta/repodb"
 	"zotregistry.io/zot/pkg/storage"
-	"zotregistry.io/zot/pkg/storage/repodb"
 )
 
 // EnableSearchExtension ...
